+++ smoke-test/smoke-spring/src/main/java/org/traffichunter/titan/smoke/springframework/SmokeSpringApplication.java
@@ -2,7 +2,7 @@ package org.traffichunter.titan.smoke.springframework;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.traffichunter.titan.springframework.stomp.EnableTitan;
+import org.traffichunter.titan.springframework.stomp.annotation.EnableTitan;
 
 @EnableTitan
 @SpringBootApplication
