+++ titan-spring-client/src/main/java/org/traffichunter/titan/springframework/stomp/annotation/TitanListener.java
@@ -21,7 +21,7 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 THE SOFTWARE.
 */
-package org.traffichunter.titan.springframework.stomp;
+package org.traffichunter.titan.springframework.stomp.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
