+++ titan-spring-client/src/main/java/org/traffichunter/titan/springframework/stomp/listener/TitanListenerEndpoint.java
@@ -1,4 +1,4 @@
-package org.traffichunter.titan.springframework.stomp;
+package org.traffichunter.titan.springframework.stomp.listener;
 
 import java.lang.reflect.Method;
